# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * noteとはてなブログの自動投稿サーバーAPI
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ArticleStatus = typeof ArticleStatus[keyof typeof ArticleStatus];


export const ArticleStatus = {
  draft: 'draft',
  queued: 'queued',
  publishing: 'publishing',
  published: 'published',
  failed: 'failed',
} as const;

export type PublishJobStatus = typeof PublishJobStatus[keyof typeof PublishJobStatus];


export const PublishJobStatus = {
  queued: 'queued',
  running: 'running',
  succeeded: 'succeeded',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export type PublishTarget = typeof PublishTarget[keyof typeof PublishTarget];


export const PublishTarget = {
  note: 'note',
  hatena: 'hatena',
} as const;

export interface Article {
  id: string;
  sourceDate: string | null;
  title: string;
  body: string;
  hashtags: string[];
  status: ArticleStatus;
  noteUrl: string | null;
  publishedUrl: string | null;
  scheduledAt: string | null;
  publishedAt: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateArticleRequest {
  /**
     * @minLength 1
     * @maxLength 200
     */
  title: string;
  /** @minLength 1 */
  body: string;
  /**
     * @maxItems 30
     * @items.maxLength 50
     */
  hashtags?: string[];
}

export interface SaveArticleRequest {
  /**
     * @minLength 10
     * @maxLength 10
     * @pattern ^[0-9]{4}-[0-9]{2}-[0-9]{2}$
     */
  date: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  title: string;
  /** @minLength 1 */
  body: string;
  image?: string | null;
}

export type SaveArticleResultStatus = typeof SaveArticleResultStatus[keyof typeof SaveArticleResultStatus];


export const SaveArticleResultStatus = {
  saved: 'saved',
} as const;

export interface SaveArticleResult {
  status: SaveArticleResultStatus;
}

export interface UpdateArticleRequest {
  /**
     * @minLength 1
     * @maxLength 200
     */
  title?: string;
  /** @minLength 1 */
  body?: string;
  /**
     * @maxItems 30
     * @items.maxLength 50
     */
  hashtags?: string[];
}

export interface PublishJob {
  id: string;
  articleId: string;
  target: PublishTarget;
  status: PublishJobStatus;
  attempts: number;
  errorMessage: string | null;
  scheduledAt: string | null;
  startedAt: string | null;
  finishedAt: string | null;
  createdAt: string;
}

export interface CreatePublishJobRequest {
  articleId: string;
  target?: PublishTarget;
  scheduledAt?: string | null;
}

/**
 * Invalid request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;

export type ListArticlesParams = {
status?: ArticleStatus;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListPublishJobsParams = {
status?: PublishJobStatus;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};