# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: noteとはてなブログの自動投稿サーバーAPI
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: articles
    description: 記事の下書き管理
  - name: compatibility
    description: Workers互換保存API
  - name: publish-jobs
    description: note / はてなブログ投稿ジョブの管理
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List note articles
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ArticleStatus"
        - name: limit
          in: query
          required: false
          schema:
            type: number
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
    post:
      operationId: createArticle
      tags: [articles]
      summary: Create a note article draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleRequest"
      responses:
        "201":
          description: Created article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "400":
          $ref: "#/components/responses/BadRequest"
  /saveArticle:
    post:
      operationId: saveArticle
      tags: [compatibility]
      summary: Save an article from a Worker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveArticleRequest"
      responses:
        "200":
          description: Article saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveArticleResult"
        "400":
          $ref: "#/components/responses/BadRequest"
  /articles/{articleId}:
    parameters:
      - $ref: "#/components/parameters/ArticleId"
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get a note article
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateArticle
      tags: [articles]
      summary: Update a note article draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArticleRequest"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteArticle
      tags: [articles]
      summary: Delete a note article
      responses:
        "204":
          description: Article deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /publish-jobs:
    get:
      operationId: listPublishJobs
      tags: [publish-jobs]
      summary: List note publishing jobs
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/PublishJobStatus"
        - name: limit
          in: query
          required: false
          schema:
            type: number
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Publishing jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublishJob"
    post:
      operationId: createPublishJob
      tags: [publish-jobs]
      summary: Queue a note article for publishing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePublishJobRequest"
      responses:
        "201":
          description: Created publishing job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishJob"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /publish-jobs/{jobId}:
    parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getPublishJob
      tags: [publish-jobs]
      summary: Get a publishing job
      responses:
        "200":
          description: Publishing job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishJob"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: cancelPublishJob
      tags: [publish-jobs]
      summary: Cancel a queued publishing job
      responses:
        "200":
          description: Cancelled publishing job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishJob"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    ArticleId:
      name: articleId
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    ArticleStatus:
      type: string
      enum: [draft, queued, publishing, published, failed]
    PublishJobStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
    PublishTarget:
      type: string
      enum: [note, hatena]
    Article:
      type: object
      properties:
        id:
          type: string
        sourceDate:
          type: string
          nullable: true
        title:
          type: string
        body:
          type: string
        hashtags:
          type: array
          items:
            type: string
        status:
          $ref: "#/components/schemas/ArticleStatus"
        noteUrl:
          type: string
          nullable: true
        publishedUrl:
          type: string
          nullable: true
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        publishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - sourceDate
        - title
        - body
        - hashtags
        - status
        - noteUrl
        - publishedUrl
        - scheduledAt
        - publishedAt
        - createdAt
        - updatedAt
    CreateArticleRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        body:
          type: string
          minLength: 1
        hashtags:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 30
      required:
        - title
        - body
    SaveArticleRequest:
      type: object
      properties:
        date:
          type: string
          minLength: 10
          maxLength: 10
          pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$"
        title:
          type: string
          minLength: 1
          maxLength: 200
        body:
          type: string
          minLength: 1
        image:
          type: string
          nullable: true
      required:
        - date
        - title
        - body
    SaveArticleResult:
      type: object
      properties:
        status:
          type: string
          enum: [saved]
      required:
        - status
    UpdateArticleRequest:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        body:
          type: string
          minLength: 1
        hashtags:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 30
    PublishJob:
      type: object
      properties:
        id:
          type: string
        articleId:
          type: string
        target:
          $ref: "#/components/schemas/PublishTarget"
        status:
          $ref: "#/components/schemas/PublishJobStatus"
        attempts:
          type: number
        errorMessage:
          type: string
          nullable: true
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - articleId
        - target
        - status
        - attempts
        - errorMessage
        - scheduledAt
        - startedAt
        - finishedAt
        - createdAt
    CreatePublishJobRequest:
      type: object
      properties:
        articleId:
          type: string
        target:
          $ref: "#/components/schemas/PublishTarget"
        scheduledAt:
          type: string
          format: date-time
          nullable: true
      required:
        - articleId