# lib/api-zod/src/generated/types/articleStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * noteとはてなブログの自動投稿サーバーAPI
 * OpenAPI spec version: 0.1.0
 */

export type ArticleStatus = typeof ArticleStatus[keyof typeof ArticleStatus];


export const ArticleStatus = {
  draft: 'draft',
  queued: 'queued',
  publishing: 'publishing',
  published: 'published',
  failed: 'failed',
} as const;